requirements.txt
astroid==1.6.4
certifi==2018.4.16
dj-database-url==0.5.0
Django==2.0.6
django-cors-headers==2.2.0
djangorestframework==3.8.2
gunicorn==19.8.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pipenv==2018.5.18
pkg-resources==0.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pylint==1.9.1
pytz==2018.4
six==1.11.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
wrapt==1.10.11
